refac: Refactored the games_teams join table
This was done to ensure that we can update the table with the new score. The old table did now have a primary key and didn't allow us to update it directly.
Editing migrations are probably the most stupid thing evar to do, so it's a NONO, but this will never go to production and if it does, it's after all the migrations will be written...